Refactors benchmarking scenarios
